google_cloud_funtions_final/requirements.txt
# Function dependencies, for example:
# package>=version

pandas
pandas-gbq
google-cloud-bigquery
google-auth==1.35.0
gcsfs
fsspec